requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
requests>=2.31.0
youtube-transcript-api>=1.0.0
yt-dlp>=2024.0
